+++ Assets/Tests/InputSystem/Plugins/InputForUITests.cs
@@ -65,7 +65,7 @@ public class InputForUITests : InputTestFixture
             EventProvider.ClearMockProvider();
         m_InputForUIEvents.Clear();
 
-        InputSystem.s_Manager.actions = storedActions;
+        InputSystem.manager.actions = storedActions;
 
 #if UNITY_EDITOR
         if (File.Exists(kAssetPath))
@@ -226,7 +226,7 @@ public class InputForUITests : InputTestFixture
             // Remove the project-wide actions asset in play mode and player.
             // It will call InputSystem.onActionChange and re-set InputSystemProvider.actionAsset
             // This the case where no project-wide actions asset is available in the project.
-            InputSystem.s_Manager.actions = null;
+            InputSystem.manager.actions = null;
         }
         Update();
 
@@ -300,7 +300,7 @@ public class InputForUITests : InputTestFixture
         Update();
         if (!useProjectWideActionsAsset)
         {
-            InputSystem.s_Manager.actions = null;
+            InputSystem.manager.actions = null;
         }
         Update();
 
@@ -337,7 +337,7 @@ public class InputForUITests : InputTestFixture
         Update();
         if (!useProjectWideActionsAsset)
         {
-            InputSystem.s_Manager.actions = null;
+            InputSystem.manager.actions = null;
         }
         Update();
 
@@ -367,7 +367,7 @@ public class InputForUITests : InputTestFixture
         Update();
         if (!useProjectWideActionsAsset)
         {
-            InputSystem.s_Manager.actions = null;
+            InputSystem.manager.actions = null;
         }
         Update();
 
@@ -419,7 +419,7 @@ public class InputForUITests : InputTestFixture
         Update();
         if (!useProjectWideActionsAsset)
         {
-            InputSystem.s_Manager.actions = null;
+            InputSystem.manager.actions = null;
         }
         Update();
 
@@ -506,7 +506,7 @@ public class InputForUITests : InputTestFixture
         Update();
         if (!useProjectWideActionsAsset)
         {
-            InputSystem.s_Manager.actions = null;
+            InputSystem.manager.actions = null;
         }
         Update();
 
@@ -561,7 +561,7 @@ public class InputForUITests : InputTestFixture
     public void DefaultActions_ShouldNotGenerateAnyVerificationWarnings(bool useProjectWideActions)
     {
         if (!useProjectWideActions)
-            InputSystem.s_Manager.actions = null;
+            InputSystem.manager.actions = null;
         Update();
         LogAssert.NoUnexpectedReceived();
     }
@@ -573,7 +573,7 @@ public class InputForUITests : InputTestFixture
         var asset = ProjectWideActionsAsset.CreateDefaultAssetAtPath(kAssetPath);
         asset.RemoveActionMap(asset.FindActionMap("UI", throwIfNotFound: true));
 
-        InputSystem.s_Manager.actions = asset;
+        InputSystem.manager.actions = asset;
         Update();
 
         LogAssert.NoUnexpectedReceived();
@@ -587,7 +587,7 @@ public class InputForUITests : InputTestFixture
         asset.RemoveActionMap(asset.FindActionMap("UI", throwIfNotFound: true));
         asset.AddActionMap(new InputActionMap("UI")); // An empty UI map should log warnings.
 
-        InputSystem.s_Manager.actions = asset;
+        InputSystem.manager.actions = asset;
         Update();
 
         var link = EditorHelpers.GetHyperlink(kAssetPath);
@@ -620,7 +620,7 @@ public class InputForUITests : InputTestFixture
         var action = asset.FindAction(actionPath);
         action.Rename("Other");
 
-        InputSystem.s_Manager.actions = asset;
+        InputSystem.manager.actions = asset;
         Update();
 
         //var link = AssetDatabase.GetAssetPath()//EditorHelpers.GetHyperlink(kAssetPath);
@@ -663,7 +663,7 @@ public class InputForUITests : InputTestFixture
 
         asset.AddActionMap(newMap);
 
-        InputSystem.s_Manager.actions = asset;
+        InputSystem.manager.actions = asset;
         Update();
 
         LogAssert.Expect(LogType.Warning, new Regex($"^InputAction with path '{actionPath}' in asset \"{kAssetPath}\" do not have any configured bindings."));
@@ -688,7 +688,7 @@ public class InputForUITests : InputTestFixture
         var expectedType = action.type;
         action.m_Type = unexpectedType; // change directly via internals for now
 
-        InputSystem.s_Manager.actions = asset;
+        InputSystem.manager.actions = asset;
         Update();
 
         LogAssert.Expect(LogType.Warning,
@@ -714,7 +714,7 @@ public class InputForUITests : InputTestFixture
         var expectedControlType = action.expectedControlType;
         action.expectedControlType = unexpectedControlType;
 
-        InputSystem.s_Manager.actions = asset;
+        InputSystem.manager.actions = asset;
         Update();
 
         LogAssert.Expect(LogType.Warning,