+++ Packages/com.unity.inputsystem/InputSystem/Actions/InputActionState.cs
@@ -1143,7 +1143,7 @@ namespace UnityEngine.InputSystem
                 "Control start index out of range");
             Debug.Assert(controlStartIndex + numControls <= totalControlCount, "Control range out of bounds");
 
-            var manager = InputSystem.s_Manager;
+            var manager = InputSystem.manager;
             for (var i = 0; i < numControls; ++i)
             {
                 var controlIndex = controlStartIndex + i;
@@ -1179,7 +1179,7 @@ namespace UnityEngine.InputSystem
                 "Control start index out of range");
             Debug.Assert(controlStartIndex + numControls <= totalControlCount, "Control range out of bounds");
 
-            var manager = InputSystem.s_Manager;
+            var manager = InputSystem.manager;
             for (var i = 0; i < numControls; ++i)
             {
                 var controlIndex = controlStartIndex + i;
@@ -1255,7 +1255,7 @@ namespace UnityEngine.InputSystem
 
             if (m_OnBeforeUpdateDelegate == null)
                 m_OnBeforeUpdateDelegate = OnBeforeInitialUpdate;
-            InputSystem.s_Manager.onBeforeUpdate += m_OnBeforeUpdateDelegate;
+            InputSystem.manager.onBeforeUpdate += m_OnBeforeUpdateDelegate;
             m_OnBeforeUpdateHooked = true;
         }
 
@@ -1264,7 +1264,7 @@ namespace UnityEngine.InputSystem
             if (!m_OnBeforeUpdateHooked)
                 return;
 
-            InputSystem.s_Manager.onBeforeUpdate -= m_OnBeforeUpdateDelegate;
+            InputSystem.manager.onBeforeUpdate -= m_OnBeforeUpdateDelegate;
             m_OnBeforeUpdateHooked = false;
         }
 
@@ -1297,7 +1297,7 @@ namespace UnityEngine.InputSystem
             // Go through all binding states and for every binding that needs an initial state check,
             // go through all bound controls and for each one that isn't in its default state, pretend
             // that the control just got actuated.
-            var manager = InputSystem.s_Manager;
+            var manager = InputSystem.manager;
             for (var bindingIndex = 0; bindingIndex < totalBindingCount; ++bindingIndex)
             {
                 ref var bindingState = ref bindingStates[bindingIndex];
@@ -1543,7 +1543,7 @@ namespace UnityEngine.InputSystem
 
                     // Check if we should suppress interaction processing notifications
                     m_Suppressed = (eventPtr != null) && eventPtr.handled &&
-                        InputSystem.s_Manager.inputEventHandledPolicy == InputEventHandledPolicy.SuppressActionEventNotifications;
+                        InputSystem.manager.inputEventHandledPolicy == InputEventHandledPolicy.SuppressActionEventNotifications;
 
                     // Check if we have multiple concurrent actuations on the same action. This may lead us
                     // to ignore certain inputs (e.g. when we get an input of lesser magnitude while already having
@@ -2132,7 +2132,7 @@ namespace UnityEngine.InputSystem
             Debug.Assert(trigger.controlIndex >= 0 && trigger.controlIndex < totalControlCount, "Control index out of range");
             Debug.Assert(trigger.interactionIndex >= 0 && trigger.interactionIndex < totalInteractionCount, "Interaction index out of range");
 
-            var manager = InputSystem.s_Manager;
+            var manager = InputSystem.manager;
             var currentTime = trigger.time;
             var control = controls[trigger.controlIndex];
             var interactionIndex = trigger.interactionIndex;
@@ -2161,7 +2161,7 @@ namespace UnityEngine.InputSystem
 
             ref var interactionState = ref interactionStates[interactionIndex];
 
-            var manager = InputSystem.s_Manager;
+            var manager = InputSystem.manager;
             manager.RemoveStateChangeMonitorTimeout(this, interactionState.timerMonitorIndex, interactionIndex);
 
             // Update state.
@@ -4313,6 +4313,19 @@ namespace UnityEngine.InputSystem
 
         internal static GlobalState s_GlobalState;
 
+        [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.BeforeSceneLoad)]
+        private static void InitializeGlobalActionState()
+        {
+            // Appears we shouldn't really reset globals in case the domain reload is enabled.
+            // This is because in that case, we've just had the whole system init'ed via static ctors
+            // Moreover, later in GlobalInialize we skip initialization specifically in this case.
+            if (!InputSystem.IsDomainReloadDisabledForPlayMode())
+                return;
+
+            ResetGlobals();
+            s_GlobalState = default;
+        }
+
         internal static ISavedState SaveAndResetState()
         {
             // Save current state
@@ -4559,40 +4572,6 @@ namespace UnityEngine.InputSystem
                         break;
                     }
                 }
-            }
-        }
-
-        internal static void DeferredResolutionOfBindings()
-        {
-            ++InputActionMap.s_DeferBindingResolution;
-            try
-            {
-                if (InputActionMap.s_NeedToResolveBindings)
-                {
-                    for (var i = 0; i < s_GlobalState.globalList.length; ++i)
-                    {
-                        var handle = s_GlobalState.globalList[i];
-
-                        var state = handle.IsAllocated ? (InputActionState)handle.Target : null;
-                        if (state == null)
-                        {
-                            // Stale entry in the list. State has already been reclaimed by GC. Remove it.
-                            if (handle.IsAllocated)
-                                s_GlobalState.globalList[i].Free();
-                            s_GlobalState.globalList.RemoveAtWithCapacity(i);
-                            --i;
-                            continue;
-                        }
-
-                        for (var n = 0; n < state.totalMapCount; ++n)
-                            state.maps[n].ResolveBindingsIfNecessary();
-                    }
-                    InputActionMap.s_NeedToResolveBindings = false;
-                }
-            }
-            finally
-            {
-                --InputActionMap.s_DeferBindingResolution;
             }
         }