+++ Packages/com.unity.inputsystem/InputSystem/Controls/InputControl.cs
@@ -83,6 +83,7 @@ namespace UnityEngine.InputSystem
     /// <seealso cref="InputDevice"/>
     /// <seealso cref="InputControlPath"/>
     /// <seealso cref="InputStateBlock"/>
+    [Serializable]
     [DebuggerDisplay("{DebuggerDisplay(),nq}")]
     public abstract class InputControl
     {
@@ -1071,7 +1072,7 @@ namespace UnityEngine.InputSystem
         private void SetOptimizedControlDataType()
         {
             // setting check need to be inline so we clear optimizations if setting is disabled after the fact
-            m_OptimizedControlDataType = InputSystem.s_Manager.optimizedControlsFeatureEnabled
+            m_OptimizedControlDataType = InputSystem.manager.optimizedControlsFeatureEnabled
                 ? CalculateOptimizedControlDataType()
                 : (FourCC)InputStateBlock.kFormatInvalid;
         }
@@ -1099,7 +1100,7 @@ namespace UnityEngine.InputSystem
         [Conditional("UNITY_EDITOR")]
         internal void EnsureOptimizationTypeHasNotChanged()
         {
-            if (!InputSystem.s_Manager.optimizedControlsFeatureEnabled)
+            if (!InputSystem.manager.optimizedControlsFeatureEnabled)
                 return;
 
             var currentOptimizedControlDataType = CalculateOptimizedControlDataType();
@@ -1283,6 +1284,7 @@ namespace UnityEngine.InputSystem
     /// <typeparam name="TValue">Type of value captured by the control. Note that this does not mean
     /// that the control has to store data in the given value format. A control that captures float
     /// values, for example, may be stored in state as byte values instead.</typeparam>
+    [Serializable]
     public abstract class InputControl<TValue> : InputControl
         where TValue : struct
     {
@@ -1326,7 +1328,7 @@ namespace UnityEngine.InputSystem
 
                 if (
                     // if feature is disabled we re-evaluate every call
-                    !InputSystem.s_Manager.readValueCachingFeatureEnabled
+                    !InputSystem.manager.readValueCachingFeatureEnabled
                     // if cached value is stale we re-evaluate and clear the flag
                     || m_CachedValueIsStale
                     // if a processor in stack needs to be re-evaluated, but unprocessedValue is still can be cached
@@ -1337,7 +1339,7 @@ namespace UnityEngine.InputSystem
                     m_CachedValueIsStale = false;
                 }
 #if DEBUG
-                else if (InputSystem.s_Manager.paranoidReadValueCachingChecksEnabled)
+                else if (InputSystem.manager.paranoidReadValueCachingChecksEnabled)
                 {
                     var oldUnprocessedValue = m_UnprocessedCachedValue;
                     var newUnprocessedValue = unprocessedValue;
@@ -1393,7 +1395,7 @@ namespace UnityEngine.InputSystem
 
                 if (
                     // if feature is disabled we re-evaluate every call
-                    !InputSystem.s_Manager.readValueCachingFeatureEnabled
+                    !InputSystem.manager.readValueCachingFeatureEnabled
                     // if cached value is stale we re-evaluate and clear the flag
                     || m_UnprocessedCachedValueIsStale
                 )
@@ -1402,7 +1404,7 @@ namespace UnityEngine.InputSystem
                     m_UnprocessedCachedValueIsStale = false;
                 }
 #if DEBUG
-                else if (InputSystem.s_Manager.paranoidReadValueCachingChecksEnabled)
+                else if (InputSystem.manager.paranoidReadValueCachingChecksEnabled)
                 {
                     var currentUnprocessedValue = ReadUnprocessedValueFromState(currentStatePtr);
                     if (CompareValue(ref currentUnprocessedValue, ref m_UnprocessedCachedValue))