+++ Packages/com.unity.inputsystem/InputSystem/Plugins/OnScreen/OnScreenStick.cs
@@ -249,16 +249,9 @@ namespace UnityEngine.InputSystem.OnScreen
             if (m_RaycastResults.Count == 0)
                 return;
 
-            var stickSelected = false;
-            foreach (var result in m_RaycastResults)
-            {
-                if (result.gameObject != gameObject) continue;
-
-                stickSelected = true;
-                break;
-            }
-
-            if (!stickSelected)
+            // Only accept selection if the top-most UI hit is within this stick's hierarchy.
+            var topResult = m_RaycastResults[0];
+            if (!topResult.gameObject.transform.IsChildOf(transform))
                 return;
 
             BeginInteraction(screenPosition, GetCameraFromCanvas());